scripts/requirements-publications.txt
PyYAML>=6.0
scholarly>=1.7.11
httpx<0.28
